InheritanceTest -> AnnotationsInheritanceTest
